# app/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "whisperX REST service",
    "description": "\n    # whisperX RESTful API\n\n    Welcome to the whisperX RESTful API! This API provides a suite of audio processing services to enhance and analyze your audio content.\n\n    ## Documentation:\n\n    For detailed information on request and response formats, consult the [WhisperX Documentation](https://github.com/m-bain/whisperX).\n\n    ## Services:\n\n    Speech-2-Text provides a suite of audio processing services to enhance and analyze your audio content. The following services are available:\n\n    1. Transcribe: Transcribe an audio/video  file into text.\n    2. Align: Align the transcript to the audio/video file.\n    3. Diarize: Diarize an audio/video file into speakers.\n    4. Combine Transcript and Diarization: Combine the transcript and diarization results.\n\n    ## Supported file extensions:\n    AUDIO_EXTENSIONS = {'.awb', '.ogg', '.mp3', '.amr', '.aac', '.m4a', '.wav', '.wma', '.oga'}\n\n    VIDEO_EXTENSIONS = {'.wmv', '.mp4', '.mov', '.avi', '.mkv'}\n\n    ",
    "version": "0.0.1"
  },
  "paths": {
...
    "/task/all": {
      "get": {
        "tags": [
          "Tasks Management"
        ],
        "summary": "Get All Tasks Status",
        "description": "Retrieve the status of all tasks.\n\nArgs:\n    service: Task management service dependency.\n\nReturns:\n    TaskListResponse: The status of all tasks.",
        "operationId": "get_all_tasks_status_task_all_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/task/{identifier}": {
      "get": {
        "tags": [
          "Tasks Management"
        ],
        "summary": "Get Transcription Status",
        "description": "Retrieve the status of a specific task by its identifier.\n\nArgs:\n    identifier (str): The identifier of the task.\n    service: Task management service dependency.\n\nReturns:\n    Result: The status of the task.\n\nRaises:\n    TaskNotFoundError: If the identifier is not found.",
        "operationId": "get_transcription_status_task__identifier__get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tasks Management"
        ],
        "summary": "Cancel Task",
        "description": "Cancel a running or queued task by its identifier.\n\nCancellation is cooperative: the background worker stops between pipeline\nstages. An active GPU step may finish before the task is fully stopped.\n\nArgs:\n    identifier (str): The identifier of the task.\n    service: Task management service dependency.\n\nReturns:\n    Response: Confirmation message of task cancellation.\n\nRaises:\n    TaskNotFoundError: If the task is not found.\n    InvalidTaskStateError: If the task is already completed, failed, or cancelled.",
        "operationId": "cancel_task_task__identifier__delete",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/audio/transcriptions": {
      "post": {
        "tags": [
          "OpenAI compatibility"
        ],
        "summary": "Transcribe audio synchronously with an OpenAI-compatible API",
        "description": "Transcribe an uploaded file and return the transcript synchronously.\n\nThe transcript is returned directly in the response. The request is also\npersisted as a task, so its result (or error) can be retrieved afterwards via\nthe `/task/{identifier}` and `/task/all` endpoints.",
        "operationId": "create_transcription_v1_audio_transcriptions_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transcript in the requested format.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/OpenAIJsonTranscriptionResponse"
                    },
                    {
                      "$ref": "#/components/schemas/OpenAIVerboseJsonResponse"
                    }
                  ]
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/x-subrip": {
                "schema": {
                  "type": "string"
                }
              },
              "text/vtt": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audio/translations": {
      "post": {
        "tags": [
          "OpenAI compatibility"
        ],
        "summary": "Translate audio to English synchronously with an OpenAI-compatible API",
        "description": "Translate an uploaded file to English and return the result synchronously.\n\nThe translation is returned directly in the response. The request is also\npersisted as a task, so its result (or error) can be retrieved afterwards via\nthe `/task/{identifier}` and `/task/all` endpoints.",
        "operationId": "create_translation_v1_audio_translations_post",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transcript in the requested format.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/OpenAIJsonTranscriptionResponse"
                    },
                    {
                      "$ref": "#/components/schemas/OpenAIVerboseJsonResponse"
                    }
                  ]
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/x-subrip": {
                "schema": {
                  "type": "string"
                }
              },
              "text/vtt": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Response": {
        "properties": {
          "identifier": {
            "type": "string",
            "title": "Identifier"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "identifier",
          "message"
        ],
        "title": "Response",
        "description": "Response model for API responses."
      },
      "Result": {
        "properties": {
          "identifier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identifier"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "partial_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Partial Text"
          },
          "partial_speaker_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Partial Speaker Count"
          },
          "partial_speakers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Partial Speakers"
          },
          "result": {
            "title": "Result"
          },
          "metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Metadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "status",
          "result",
          "error"
        ],
        "title": "Result",
        "description": "Model for a result with status, result data, metadata, and optional error."
      },
...
      "TaskListResponse": {
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskSummaryResponse"
            },
            "type": "array",
            "title": "Tasks",
            "description": "List of task summaries"
          }
        },
        "type": "object",
        "required": [
          "tasks"
        ],
        "title": "TaskListResponse",
        "description": "DTO for returning a list of task summaries."
      },
      "TaskSummaryResponse": {
        "properties": {
          "identifier": {
            "type": "string",
            "title": "Identifier",
            "description": "Unique identifier for the task"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Current status of the task"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type",
            "description": "Type of task"
          },
          "file_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Name",
            "description": "Name of the file"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url",
            "description": "URL of the file"
          },
          "audio_duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Duration",
            "description": "Duration of the audio"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "description": "Language of the audio"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error message if task failed"
          },
          "duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration",
            "description": "Execution duration in seconds"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time",
            "description": "Task start time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time",
            "description": "Task end time"
          }
        },
        "type": "object",
        "required": [
          "identifier",
          "status",
          "task_type"
        ],
        "title": "TaskSummaryResponse",
        "description": "DTO for returning task summary in list operations.\n\nUsed for GET /task/all endpoint to provide a lighter response."
      },
...
    }
  },
...
}